requirements.txt
packaging==22.0
Pillow==9.3.0
pip==22.3.1
pytesseract==0.3.10
setuptools==60.2.0
wheel==0.37.1
